Chemical-Model is working on a basic level

+++ CBRN_Project/CBRN_Project/MVVM/Models/Input/Vehicle Shelter/BlastShieldingProtection.cs
@@ -9,7 +9,7 @@ namespace CBRN_Project.MVVM.Models
     
     public class BlastShieldingProtection
     {
-        public float ProtectionFactor { get; set; } = 1;
+        public double ProtectionFactor { get; set; } = 1;
 
         public BlastShieldingProtection()
         {
